requirements.txt
ccxt==4.2.20
pydantic==2.5.3
pandas==2.2.0
ta==0.11.0
boto3 --no-cache-dir
aiobotocore==1.3.3
scipy==1.9.0 